added for storage
This is a JavaScript code that receives data from connected gamepads to the device via the USB port. The code outputs the data received from the gamepad to an HTML document, the device name if any, the buttons and their state (pressed/not pressed), as well as the axes in the form of converted numbers and visual indicators of the axis positions. The code contains a small demo to understand how the code works. This project is a tool for manually adjusting the joystick axes when installing joysticks on gamepad-type devices. This code shows errors in rotation and deviations when tightening the fasteners on the joysticks during fixation on the gamepad body. The code does not provide 100% installation accuracy, but allows you to minimize material stress when installing devices and securing them. There are no instructions for working with the tool, they were invented in the course of working with the tool specifically in each case during installations, in general, just combine all the axes along the guides, fix the joystick, correct the rotation and tension of the rubber bands, etc., restore the zeros along the axes using the guides and fix it with glue so that it does not run away. Good luck to anyone who needs it, take it, use it, write reviews, p.s.: the code was written by Grok, I only edited it for my tasks :)